Run for multiple test cases, set up a framework
